made NOLF2 const char* correct, and updated FontMgr to follow suit
and extended object life time for a few temporaries.
libCShell.so builds when treating warnings as errors
(only the deprecation warning silenced)

+++ NOLF2/ClientShellDLL/ClientShellShared/NodeController.cpp
@@ -105,7 +105,7 @@ LTBOOL CNodeController::Init(CCharacterFX* pCharacterFX)
 		{
 			_ASSERT(eModelNode < kMaxNodes);
 
-			if (eModelNode < kMaxNodes)
+			if ((uint)eModelNode < (uint)kMaxNodes)
 			{
 				m_aNodes[eModelNode].eModelNode = eModelNode;
 				m_aNodes[eModelNode].hModelNode = hCurNode;
