+++ NOLF2/ClientShellDLL/TO2/ScreenHost.h
@@ -39,7 +39,7 @@ public:
 	void	ReadyLaunch(LTBOOL bReady);
 
 protected:
-	void	HandleCallback(uint32 dwParam1, uint32 dwParam2);
+	void	HandleCallback(uintptr_t dwParam1, uint32 dwParam2);
 	void	CreateDefaultCampaign();
 
     LTBOOL	LaunchGame();
@@ -53,7 +53,7 @@ protected:
 
 
 protected:
-    uint32  OnCommand(uint32 dwCommand, uint32 dwParam1, uint32 dwParam2);
+    uint32  OnCommand(uint32 dwCommand, uintptr_t dwParam1, uint32 dwParam2);
 	CString	m_sSessionName;
 	CString	m_sCampaignName;
 	CString	m_sPassword;
