+++ NOLF2/ClientShellDLL/TO2/ScreenHostWeapons.h
@@ -39,7 +39,7 @@ protected:
 		kGearType,
 	};
 
-    uint32  OnCommand(uint32 dwCommand, uint32 dwParam1, uint32 dwParam2);
+    uint32  OnCommand(uint32 dwCommand, uintptr_t dwParam1, uint32 dwParam2);
     LTBOOL  FillAvailList();
 	void	LoadItemList();
 	void	MakeDefaultItemList();
