+++ NOLF2/ClientShellDLL/TO2/ScreenJoinLAN.h
@@ -36,7 +36,7 @@ public:
 	void	DoJoin();
 
 protected:
-    uint32  OnCommand(uint32 dwCommand, uint32 dwParam1, uint32 dwParam2);
+    uint32  OnCommand(uint32 dwCommand, uintptr_t dwParam1, uint32 dwParam2);
 
 	LTBOOL	InitSessions();
 	void	FindServers();
