+++ NOLF2/ClientShellDLL/TO2/ScreenMulti.h
@@ -31,8 +31,8 @@ public:
     void    OnFocus(LTBOOL bFocus);
 
 protected:
-    uint32	OnCommand(uint32 dwCommand, uint32 dwParam1, uint32 dwParam2);
-	uint32	HandleCallback(uint32 dwParam1, uint32 dwParam2);
+    uint32	OnCommand(uint32 dwCommand, uintptr_t dwParam1, uint32 dwParam2);
+	uint32	HandleCallback(uintptr_t dwParam1, uint32 dwParam2);
 	void	ChangeCDKey();
 	void	RequestMOTD();
 	void	RequestValidate();
