+++ NOLF2/ClientShellDLL/TO2/ScreenPlayer.h
@@ -40,13 +40,13 @@ public:
 	void	PrevModel();
 
 protected:
-	void	HandleCallback(uint32 dwParam1, uint32 dwParam2);
+	void	HandleCallback(uintptr_t dwParam1, uint32 dwParam2);
 
 	void	UpdateBandwidth();
 
 	void	UpdateChar();
 
-    uint32  OnCommand(uint32 dwCommand, uint32 dwParam1, uint32 dwParam2);
+    uint32  OnCommand(uint32 dwCommand, uintptr_t dwParam1, uint32 dwParam2);
 
 	CLTGUIColumnCtrl*	m_pName;
 	CLTGUITextCtrl*		m_pModel;
