+++ NOLF2/ClientShellDLL/TO2/ScreenSave.h
@@ -31,7 +31,7 @@ public:
 	LTBOOL SaveGame(uint32 slot);
 
 protected:
-    uint32  OnCommand(uint32 dwCommand, uint32 dwParam1, uint32 dwParam2);
+    uint32  OnCommand(uint32 dwCommand, uintptr_t dwParam1, uint32 dwParam2);
 	void	BuildSavedLevelList();
 	void	ClearSavedLevelList();
